Remove test for the product component.
I removed tests for product as I got an error specifying useNavigator may
only be used in the context of the <Router> component.

Even though, the product component is in the context of the <Router>
component. I think jest is not aware of this fact. But I moved the tests
to shop and everything works fine so far.