Return a copy of cart instead of cart itself.
I return a copy of cart instead of cart itself because React uses
Object.is() to determine if state has changed. Object.is() will compare
the reference of an array but not its content meaning that in React's
eyes nothing has changed even though the contents of cart would have
benn updated.