Store information about the product
I stored information about the product in localStorage instead of an
instance. The reason is, an instance of a product is an object that
contains functions and functions cannot be stringified since they are
not JSON. To get around this issue, I just store product data and then
retrieve it again to create new product instances (logic which is
handled in the shop module).