Update cart regardless of what addProduct returns.
I now update the cart even if addProduct returns false because sometimes
when the user adds a product to cart that is already in cart, addProduct
will return false as it won't add any new product but update the
quantity of the ones that are already added. So, only updating cart when
addProduct returns true, prevents the number of items in cart updating
correctly.